backend/requirements.txt
annotated-doc
annotated-types
anyio
authlib
bcrypt
certifi
cffi
click
cloudinary
colorama
cryptography
dnspython
ecdsa
email-validator
fastapi
h11
httpcore
httpx
idna
itsdangerous
motor
passlib
pydantic
pydantic-settings
pydantic_core
pyasn1
pycparser
pymongo
python-dotenv
python-jose
python-multipart
rsa
six
starlette
typing-extensions
typing-inspection
urllib3
uvicorn